fix types

+++ pyquil/api/_qpu.py
@@ -118,7 +118,7 @@ class QPU(QAM[QPUExecuteResponse]):
         *,
         quantum_processor_id: str,
         priority: int = 1,
-        timeout: Optional[int] = 30,
+        timeout: Optional[float] = 30.0,
         client_configuration: Optional[QCSClient] = None,
         endpoint_id: Optional[str] = None,
         execution_options: Optional[ExecutionOptions] = None,
@@ -144,7 +144,7 @@ class QPU(QAM[QPUExecuteResponse]):
         self._quantum_processor_id = quantum_processor_id
         if execution_options is None:
             execution_options_builder = ExecutionOptionsBuilder.default()
-            execution_options_builder.timeout_seconds = int(timeout)
+            execution_options_builder.timeout_seconds = timeout
             execution_options_builder.connection_strategy = ConnectionStrategy.default()
             if endpoint_id is not None:
                 execution_options_builder.connection_strategy = ConnectionStrategy.endpoint_id(endpoint_id)
